+++ packages/framework/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @meridianjs/framework
 
+## 1.6.0
+
+### Patch Changes
+
+- @meridianjs/types@1.6.0
+- @meridianjs/framework-utils@1.6.0
+
 ## 1.5.0
 
 ### Patch Changes