+++ packages/modules/auth/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @meridianjs/auth
 
+## 1.6.0
+
+### Minor Changes
+
+- Added private workspaces, and workspace access check fix
+
+### Patch Changes
+
+- @meridianjs/types@1.6.0
+- @meridianjs/framework-utils@1.6.0
+
 ## 1.5.0
 
 ### Patch Changes